Moved tests in the test folder
