+++ tests/hooks/useDataProvider.test.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import {act} from 'react-dom/test-utils';
 import configureMockStore from 'redux-mock-store';
-import useDataProvider, {IDataProviderConfig} from './useDataProvider';
-import prepareMiddleware from '../../tests/storeMiddlewareMock';
-import mountWithApp from '../../tests/mountWithApp';
+import useDataProvider, {IDataProviderConfig} from '../../src/hooks/useDataProvider';
+import prepareMiddleware from '../storeMiddlewareMock';
+import mountWithApp from '../mountWithApp';
 
 const mockStore = configureMockStore([prepareMiddleware]);
