+++ tests/hooks/useDataSelect.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {act} from 'react-dom/test-utils';
-import mountWithApp from '../../tests/mountWithApp';
-import useDataSelect, {IDataSelectConfig} from './useDataSelect';
+import mountWithApp from '../mountWithApp';
+import useDataSelect, {IDataSelectConfig} from '../../src/hooks/useDataSelect';
 
 const MockResultComponent = (props: any) => (
     <div className='list'>
